+++ plugin/uninstall.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+echo "Deleting /lib/systemd/system/touch_display_lite.service"
+rm /lib/systemd/system/digiplayer.service
+
+echo "Done"
+echo "pluginuninstallend"